refactor(app): break the paper_report/paper_trading cycle
`paper_report` imported thirteen items from `paper_trading` while
`paper_trading` re-exported `HistoryRow`, `LedgerAction` and `LedgerScope`
back, so the reading half and the writing half of paper trading could
neither be read, moved nor tested apart. PR #282 split them and welded
them in the same change; nothing in the repository saw it.

The shared presentation vocabulary moves down into `paper_chrome`, a
neutral module that imports neither half — `PositionSummary`, `caption`,
`pill_toggle`, `points_color`, `position_word`, the four `fmt_*`
formatters, `sanitize_symbol` and `list_symbol_folders`. The same move
`plot_area` was out of `pane`.

`fmt_offset_minute` and `today` land in `paper_calendar` rather than in
`paper_chrome`: both need `CivilDate` and `civil_utc`, which
`paper_calendar` owns, and `paper_calendar` paints its day cells with
`points_color` and `fmt_signed_points`. Hosting all thirteen in one module
would have traded this cycle for an identical one between the new module
and the calendar.

The `HistoryRow` / `LedgerAction` / `LedgerScope` re-export stays: it is a
deliberate facade with callers in `app.rs` and `dock.rs`, and it is
one-directional now that the other edge is gone. Every caller of a moved
helper is repointed at its new home, so `paper_calendar` and `paper_home`
no longer name `paper_trading` for a formatter either.

Behaviour is unchanged and the tests travelled with the items they cover:
`the_report_numbers_are_fixed`, the byte-for-byte pin on a whole report,
is untouched and passes.
