+++ crates/app/src/main.rs
@@ -54,6 +54,7 @@ mod orderflow_view;
 mod orderflow_worker;
 mod pane;
 mod paper_calendar;
+mod paper_chrome;
 mod paper_home;
 mod paper_hud;
 mod paper_report;
